Update Direct Lighting
In this tutorial we will be adding normal to the Mesh class that we created in previous tutorials, then
drawing it using a simple Phong diffuse shader. Then we will use the same shader to draw an
OBJMesh and extend the shader to incorporate a specular component and use the OBJ’s materials.